pyproject.toml
[project]
name = "RMT_AO3_RSS"
version = "1.0.0"
description = "RSS to Telegram Bot for Archive of Our Own"
authors = [
    {name = "RMT AO3 RSS Bot", email = "bot@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.10",
    "python-telegram-bot>=20.7",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
rss-bot = "bot:main"
rss-bot-test = "test_bot:test_bot"
rss-bot-run = "run_bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
